Move `python/src/ir.h` to `triton/Tools/OpBuilder.h`
Using `python/src/ir.h` in `PluginUtils.h` causes build failures for
plugins (e.g., this triton-ext [failure]) because those headers are
not included in the C++ artifact package plugins build against
(introduced in #9534). And they should not be: the C++ source in the
`python` directory is used for building the Python-interfacing library.
This change fixes all of that by moving the header-defined
`TritonOpBuilder` class to the `include` directory.

[failure]: https://github.com/triton-lang/triton-ext/actions/runs/23460115506/job/68259370483#step:12:86
